refactor(promote): split into to-spec/ and to-pr/ subdirs
promote/ now mirrors the two promotion pipelines:

  vsdd/promote/
  ├── to-spec/                    goal → tech-spec decomposition
  │   ├── cardinality.ts          ← was phase-1c/cardinality.ts
  │   ├── decompose.ts
  │   ├── decompose.mustache
  │   ├── spec.ts
  │   └── spec.mustache
  └── to-pr/                      tech-spec → PR (gates + prompts)
      ├── gates.ts
      ├── red.prompt.md
      ├── green.prompt.md
      └── green-no-runner.prompt.md

cardinality.ts moved from phase-1c/ into to-spec/ — it feeds the
decompose agent's prompt, not Phase 1c itself. phase-1c/ dir removed
(was empty after budget.ts moved to vsdd/budget.ts earlier).

All imports + workflow references updated. deno check clean.
