+++ .github/scripts/vsdd/promote/to-spec/decompose.mustache
