+++ .github/scripts/vsdd/promote/to-spec/spec.mustache
